requirements.txt
discord.py>=2.3.0
requests>=2.31.0
spotipy>=2.23.0
pyotp>=2.9.0
python-dotenv>=1.0.0
flask>=2.3.0

